add_float_range_column() and float_range_column()

+++ ranges.py
@@ -1,10 +1,11 @@
 import duckdb as d
-from math import ceil
+from math import ceil, floor
 # import pandas as pd
 # import polars as pl
 # from typing import Generator
 from . import cx, tn, tn_prefix, tn_sep, tn_gen
 from . import add_column, replace_column, drop_column, set_type
+from IPython.display import display
 
 def add_integer_range_column(object, basecolumn, new_colname, minvalue=0, maxvalue=None, numranges=None, rangesize=10, onlystart=False):
   """
@@ -102,6 +103,70 @@ def age_range_column(object, column, minage=0, maxage=None, numranges=None, ager
     base_column=f"{column}"
     new_colname=f"{column}_range"
     table=add_age_range_column(object,base_column,new_colname, minage=minage,maxage=maxage,numranges=numranges,agerange=agerange,onlystart=onlystart,onlyadults=onlyadults,adultage=adultage)
+    if onlystart:
+      table=replace_column(table,base_column,new_colname)
+      table=drop_column(table,new_colname)
+    else:
+      table=drop_column(table,base_column)
+    return table
+  return object
+
+def add_float_range_column(object, basecolumn, new_colname, minvalue=0, maxvalue=10, numranges=None, rangesize=1, decimals=1, onlystart=False):
+  """
+  Add range column to object, basing its values on basecolumn.
+  Will round to n-decimals decimals
+  """
+  if isinstance(object, d.duckdb.DuckDBPyRelation) \
+    and isinstance(basecolumn,str) \
+    and isinstance(new_colname,str):
+    table_name=next(tn_gen)
+    table=object.to_table(table_name)
+    if onlystart:
+      add_col_str=f"ALTER TABLE {table_name} ADD COLUMN {new_colname} DECIMAL(18,{decimals})"
+    else:
+      add_col_str=f"ALTER TABLE {table_name} ADD COLUMN {new_colname} DECIMAL(18,{decimals})[]"
+    d.sql(add_col_str)
+    table=d.sql(f"FROM {table_name}")
+    # display(table.limit(10))
+    if numranges:
+      # Create the ranges
+      if maxvalue:
+        total_range=maxvalue-minvalue
+      else:
+        max_val_str=f"SELECT MAX({basecolumn}) FROM object"
+        max_val_tuple=d.sql(max_val_str).fetchone()
+        maxvalue=max_val_tuple[0]
+      rangesize=(maxvalue-minvalue)/numranges
+      if onlystart:
+        upd_col_str=f"UPDATE {table_name} SET {new_colname}=FLOOR((FLOOR({basecolumn}/(({maxvalue}-{minvalue})/{numranges}))*(({maxvalue}-{minvalue})/{numranges}))*(10**{decimals}))/(10**{decimals})"
+      else:
+        rstart_str=f"FLOOR((FLOOR({basecolumn}/(({maxvalue}-{minvalue})/{numranges}))*(({maxvalue}-{minvalue})/{numranges}))*(10**{decimals}))/(10**{decimals})"
+        rend_str=f"{rangesize}+FLOOR((FLOOR({basecolumn}/(({maxvalue}-{minvalue})/{numranges}))*(({maxvalue}-{minvalue})/{numranges}))*(10**{decimals}))/(10**{decimals})-(1/10**{decimals})"
+        upd_col_str=f"UPDATE {table_name} SET {new_colname}=[{rstart_str},{rend_str}]"
+    else:
+      # The range size must be given
+      numranges=ceil((maxvalue-minvalue)/rangesize)
+      if onlystart:
+        upd_col_str=f"UPDATE {table_name} SET {new_colname}=FLOOR((FLOOR({basecolumn}/(({maxvalue}-{minvalue})/{numranges}))*(({maxvalue}-{minvalue})/{numranges}))*(10**{decimals}))/(10**{decimals})"
+      else:
+        rstart_str=f"FLOOR((FLOOR({basecolumn}/(({maxvalue}-{minvalue})/{numranges}))*(({maxvalue}-{minvalue})/{numranges}))*(10**{decimals}))/(10**{decimals})"
+        rend_str=f"{rangesize}+FLOOR((FLOOR({basecolumn}/(({maxvalue}-{minvalue})/{numranges}))*(({maxvalue}-{minvalue})/{numranges}))*(10**{decimals}))/(10**{decimals})-(1/10**{decimals})"
+        upd_col_str=f"UPDATE {table_name} SET {new_colname}=[{rstart_str},{rend_str}]"
+    d.sql(upd_col_str)
+    table=d.sql(f"FROM {table_name}")
+    return table
+  return object
+
+def float_range_column(object, column, minvalue=0, maxvalue=0, numranges=None, rangesize=1, decimals=1, onlystart=False):
+  """
+  Replace all values in selected column with range values. column must be numeric, and number of ranges must be integer
+  If onlystart==True, it replaces each value with the value at start of corresponding range.
+  """
+  if isinstance(object, d.duckdb.DuckDBPyRelation) \
+    and isinstance(column,str):
+    base_column=f"{column}"
+    new_colname=f"{column}_range"
+    table=add_float_range_column(object,base_column,new_colname, minvalue=minvalue,maxvalue=maxvalue,numranges=numranges,rangesize=rangesize, onlystart=onlystart)
     if onlystart:
       table=replace_column(table,base_column,new_colname)
       table=drop_column(table,new_colname)
